requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# LangGraph for agents
langgraph>=0.0.20

# Ollama
ollama>=0.1.0

# Vector store and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Finance data
yfinance>=0.2.0

# News API
worldnewsapi>=1.3.2

# Environment management
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0